LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-import-inner-scope.tentative-expected.txt

Inside
Outside

FAIL @scope within scope-imported stylesheet matches assert_equals: expected "1" but got ""
PASS @scope within scope-imported does not ignore import scope
